+++ R/globals.R
@@ -1,2 +1,2 @@
 # Declare global variables to avoid R CMD check NOTEs
-utils::globalVariables(c("cheysson_palettes", "cheysson_patterns"))
+utils::globalVariables(c("cheysson_palettes", "cheysson_patterns", "cheysson_labels"))